Backend/requirements.txt
Django==5.0
psycopg2-binary==2.9.11
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-decouple==3.8
pgvector==0.4.2
numpy==2.4.1
dj-database-url==2.1.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
djangorestframework-simplejwt==5.3.2
drf-spectacular==0.27.0
django-filter==24.1
sentence-transformers==3.0.0
